Refactor tab components
